requirements.txt
django==1.11.5
django-bootstrap4==0.0.1
django-easy-pdf>=0.2.0.dev1
django-pyodbc-azure>=1.11.0
gunicorn==19.6.0
psycopg2==2.6.2
pyodbc==4.0.17
whitenoise==3.3.0
WeasyPrint>=0.34
